Add url length checks in image_storing functions

+++ graph/graph_api/apis/image_storing.py
@@ -27,20 +27,21 @@ def get_data_from_gcs(bucket_name, file_url):
         -bucket_name = the name of the GCP bucket where to look for the given file.
         -file_url = url for the file stored in GCP bucket that needs to be retrieved. 
     Return:
-        A byte object that represents the retrieved file or None if an error occurs.
+        A byte object that represents the retrieved file or None if an error occurs or file_url is of length 0.
     Raise:
         A credentials related exceptions or an exception if the file is not found.
     '''
-    try:
-        client = storage.Client()#This will look for a GOOGLE_APPLICATION_CREDENTIALS env variable.
-        bucket = client.bucket(bucket_name)
-        target_key = get_file_name_from_url(file_url)
-        blob = bucket.get_blob(target_key).download_as_string()# get bucket data as blob
-        json_data = blob.download_as_string()# Download the contents of this blob as a bytes object.
-        return json_data
-    
-    except Exception as e:
-        print(e)
+    if(len(file_url) > 0):
+        try:
+            client = storage.Client()#This will look for a GOOGLE_APPLICATION_CREDENTIALS env variable.
+            bucket = client.bucket(bucket_name)
+            target_key = get_file_name_from_url(file_url)
+            blob = bucket.get_blob(target_key).download_as_string()# get bucket data as blob
+            json_data = blob.download_as_string()# Download the contents of this blob as a bytes object.
+            return json_data
+        
+        except Exception as e:
+            print(e)
 
     return None
 
@@ -52,13 +53,14 @@ def delete_data_from_gcs(bucket_name, file_url):
     Raise:
         A credentials related exceptions or an exception if the file is not found.
     '''
-    try:
-        client = storage.Client()#This will look for a GOOGLE_APPLICATION_CREDENTIALS env variable.
-        bucket = client.bucket(bucket_name)
-        target_key = get_file_name_from_url(file_url)
-        bucket.delete_blob(target_key)
-    except Exception as e:
-        print(e)
+    if(len(file_url) > 0):
+        try:
+            client = storage.Client()#This will look for a GOOGLE_APPLICATION_CREDENTIALS env variable.
+            bucket = client.bucket(bucket_name)
+            target_key = get_file_name_from_url(file_url)
+            bucket.delete_blob(target_key)
+        except Exception as e:
+            print(e)
 
 def update_data_from_gcs(bucket_name, old_file_url, new_data):
     '''Function that update an image in a given bucket in GCP and returns its new url.
@@ -75,8 +77,9 @@ def update_data_from_gcs(bucket_name, old_file_url, new_data):
         client = storage.Client()#This will look for a GOOGLE_APPLICATION_CREDENTIALS env variable.
         bucket = client.bucket(bucket_name)
         #delete the previously stored file.
-        old_target_key = get_file_name_from_url(old_file_url)
-        bucket.delete_blob(old_target_key)
+        if(len(old_file_url) > 0):
+            old_target_key = get_file_name_from_url(old_file_url)
+            bucket.delete_blob(old_target_key)
         #post the new file.
         new_target_key = str(uuid.uuid4())# unique file name in the bucket.
         bucket.blob(new_target_key).upload_from_string(new_data)
@@ -84,6 +87,8 @@ def update_data_from_gcs(bucket_name, old_file_url, new_data):
     
     except Exception as e:
         print(e)
+    
+    return None
 
 def get_file_name_from_url(file_url):
     return str.split("/")[-1]
